refactor(aipanel): 拆出会话首页与导航 hook
- 新增 components/ai/useAIConversationHome.ts：会话列表与供应商状态/派生、首页数据刷新、恶魔模式开关、终端标签映射、会话打开/回首页/恢复备份/重命名/删除、供应商切换与 diff 汇总入口
- AIPanel.tsx 3769 -> 3099 行
